study file input and mybatis selectkey

+++ src/main/java/com/moveone/app/regions/RegionDTO.java
@@ -6,13 +6,13 @@ public class RegionDTO {
 	//3. 생성자는 여러개 가능 하지만 기본 생성자는 필수
 	//4. 멤버변수의 이름과 Datatype은 DB Table의 column명과 Datatype과 일치
 	
-	private Integer region_id;
+	private Long region_id;
 	private String region_name;
 	
-	public Integer getRegion_id() {
+	public Long getRegion_id() {
 		return region_id;
 	}
-	public void setRegion_id(Integer region_id) {
+	public void setRegion_id(Long region_id) {
 		this.region_id = region_id;
 	}
 	public String getRegion_name() {
